+++ src/index.js
@@ -19,7 +19,7 @@ import rules from "./build/rules.js";
 const plugin = {
 	meta: {
 		name: "@eslint/json",
-		version: "0.13.1", // x-release-please-version
+		version: "0.14.0", // x-release-please-version
 	},
 	languages: {
 		json: new JSONLanguage({ mode: "json" }),
@@ -29,6 +29,7 @@ const plugin = {
 	rules,
 	configs: {
 		recommended: {
+			name: "@eslint/json/recommended",
 			plugins: {},
 			rules: recommendedRules,
 		},