+++ tests/rules/no-duplicate-keys.test.js
@@ -31,6 +31,10 @@ ruleTester.run("no-duplicate-keys", rule, {
 		'{"foo": 1, "bar": {"bar": 2}}',
 		'{"foo": { "bar": 5 }, "bar": 6 }',
 		{
+			code: '{"foo": 1, "bar": {"bar": 2}}',
+			language: "json/jsonc",
+		},
+		{
 			code: "{foo: 1, bar: {bar: 2}}",
 			language: "json/json5",
 		},
@@ -64,6 +68,20 @@ ruleTester.run("no-duplicate-keys", rule, {
 					column: 5,
 					endLine: 5,
 					endColumn: 10,
+				},
+			],
+		},
+		{
+			code: '{"foo": 1, "foo": 2}',
+			language: "json/jsonc",
+			errors: [
+				{
+					messageId: "duplicateKey",
+					data: { key: "foo" },
+					line: 1,
+					column: 12,
+					endLine: 1,
+					endColumn: 17,
 				},
 			],
 		},
@@ -134,11 +152,66 @@ ruleTester.run("no-duplicate-keys", rule, {
 			],
 		},
 		{
+			code: '{"foot": 1, "fo\\u006ft": 2}',
+			errors: [
+				{
+					messageId: "duplicateKey",
+					data: { key: "fo\\u006ft" },
+					line: 1,
+					column: 13,
+					endLine: 1,
+					endColumn: 24,
+				},
+			],
+		},
+		{
+			code: '{"foot": 1, "fo\\u006ft": 2}',
+			language: "json/jsonc",
+			errors: [
+				{
+					messageId: "duplicateKey",
+					data: { key: "fo\\u006ft" },
+					line: 1,
+					column: 13,
+					endLine: 1,
+					endColumn: 24,
+				},
+			],
+		},
+		{
+			code: '{"foot": 1, "fo\\u006ft": 2}',
+			language: "json/json5",
+			errors: [
+				{
+					messageId: "duplicateKey",
+					data: { key: "fo\\u006ft" },
+					line: 1,
+					column: 13,
+					endLine: 1,
+					endColumn: 24,
+				},
+			],
+		},
+		{
+			code: "{foot: 1, fo\\u006ft: 2}",
+			language: "json/json5",
+			errors: [
+				{
+					messageId: "duplicateKey",
+					data: { key: "fo\\u006ft" },
+					line: 1,
+					column: 11,
+					endLine: 1,
+					endColumn: 20,
+				},
+			],
+		},
+		{
 			code: '{"f\\u006fot": 1, "fo\\u006ft": 2}',
 			errors: [
 				{
 					messageId: "duplicateKey",
-					data: { key: "foot" },
+					data: { key: "fo\\u006ft" },
 					line: 1,
 					column: 18,
 					endLine: 1,
@@ -152,7 +225,7 @@ ruleTester.run("no-duplicate-keys", rule, {
 			errors: [
 				{
 					messageId: "duplicateKey",
-					data: { key: "foot" },
+					data: { key: "fo\\u006ft" },
 					line: 1,
 					column: 18,
 					endLine: 1,
@@ -166,7 +239,7 @@ ruleTester.run("no-duplicate-keys", rule, {
 			errors: [
 				{
 					messageId: "duplicateKey",
-					data: { key: "foot" },
+					data: { key: "fo\\u006ft" },
 					line: 1,
 					column: 18,
 					endLine: 1,
@@ -180,7 +253,7 @@ ruleTester.run("no-duplicate-keys", rule, {
 			errors: [
 				{
 					messageId: "duplicateKey",
-					data: { key: "foot" },
+					data: { key: "fo\\u006ft" },
 					line: 1,
 					column: 16,
 					endLine: 1,