+++ tests/rules/no-empty-keys.test.js
@@ -27,6 +27,10 @@ ruleTester.run("no-empty-keys", rule, {
 		'{"foo": 1, "bar": 2}',
 		{
 			code: '{"foo": 1, "bar": 2, "baz": 3}',
+			language: "json/jsonc",
+		},
+		{
+			code: '{"foo": 1, "bar": 2, "baz": 3}',
 			language: "json/json5",
 		},
 		{
@@ -49,6 +53,32 @@ ruleTester.run("no-empty-keys", rule, {
 		},
 		{
 			code: '{"  ": 1}',
+			errors: [
+				{
+					messageId: "emptyKey",
+					line: 1,
+					column: 2,
+					endLine: 1,
+					endColumn: 6,
+				},
+			],
+		},
+		{
+			code: '{"": 1}',
+			language: "json/jsonc",
+			errors: [
+				{
+					messageId: "emptyKey",
+					line: 1,
+					column: 2,
+					endLine: 1,
+					endColumn: 4,
+				},
+			],
+		},
+		{
+			code: '{"  ": 1}',
+			language: "json/jsonc",
 			errors: [
 				{
 					messageId: "emptyKey",